improvement algs + exec time

+++ fib_algs.h
@@ -1,9 +1,9 @@
 #ifndef APA_WORK_FIB_ALGS_H
 #define APA_WORK_FIB_ALGS_H
 
-extern long int iteration_recursive;
-extern long int iteration_iterative;
-extern long int iteration_binary;
+extern long long iteration_recursive;
+extern long long iteration_iterative;
+extern long long iteration_binary;
 
 long long fib_recursive(int n);
